+++ meson.build
@@ -811,6 +811,7 @@ if build_xserver
         libxserver_present,
         libxserver_xres,
         libxserver_xext,
+        libxserver_security,
         libxserver_miext_sync,
         libxserver_xfixes,
         libxserver_xi,
